Pull out dentry walker to FileEvents.h
